+++ lib/rspec-html-matchers.rb
@@ -1,3 +1,4 @@
+# encoding: UTF-8
 require 'nokogiri'
 
 module RSpec
@@ -9,9 +10,11 @@ module RSpec
       attr_reader :failure_message, :negative_failure_message
       attr_reader :parent_scope, :current_scope
 
-      TAG_NOT_FOUND_MSG        = %Q|expected following:\n%s\nto have at least 1 element matching "%s", found 0.|
+      TAG_FOUND_DESC           = %Q|have at least 1 element matching "%s"|
+      TAG_NOT_FOUND_MSG        = %Q|expected following:\n%s\nto #{TAG_FOUND_DESC}, found 0.|
       TAG_FOUND_MSG            = %Q|expected following:\n%s\nto NOT have element matching "%s", found %s.|
-      WRONG_COUNT_MSG          = %Q|expected following:\n%s\nto have %s element(s) matching "%s", found %s.|
+      COUNT_DESC               = %Q|have %s element(s) matching "%s"|
+      WRONG_COUNT_MSG          = %Q|expected following:\n%s\nto #{COUNT_DESC}, found %s.|
       RIGHT_COUNT_MSG          = %Q|expected following:\n%s\nto NOT have %s element(s) matching "%s", but found.|
       BETWEEN_COUNT_MSG        = %Q|expected following:\n%s\nto have at least %s and at most %s element(s) matching "%s", found %s.|
       RIGHT_BETWEEN_COUNT_MSG  = %Q|expected following:\n%s\nto NOT have at least %s and at most %s element(s) matching "%s", but found %s.|
@@ -71,6 +74,15 @@ module RSpec
           true
         else
           false
+        end
+      end
+
+      def description
+        # TODO should it be more complicated?
+        if @options.has_key?(:count)
+          COUNT_DESC % [@options[:count],@tag]
+        else
+          TAG_FOUND_DESC % @tag
         end
       end
 
@@ -244,7 +256,7 @@ module RSpec
     def have_form action_url, method, options={}, &block
       options[:with] ||= {}
       id = options[:with].delete(:id)
-      tag = 'form'; tag += '#'+id if id
+      tag = 'form'; tag << '#'+id if id
       options[:with].merge!(:action => action_url)
       options[:with].merge!(:method => method.to_s)
       have_tag tag, options, &block
@@ -397,7 +409,7 @@ module RSpec
     def with_select name, options={}, &block
       options[:with] ||= {}
       id = options[:with].delete(:id)
-      tag='select'; tag += '#'+id if id
+      tag='select'; tag << '#'+id if id
       options[:with].merge!(:name => name)
       @__current_scope_for_nokogiri_matcher.should have_tag(tag, options, &block)
     end
@@ -405,7 +417,7 @@ module RSpec
     def without_select name, options={}, &block
       options[:with] ||= {}
       id = options[:with].delete(:id)
-      tag='select'; tag += '#'+id if id
+      tag='select'; tag << '#'+id if id
       options[:with].merge!(:name => name)
       @__current_scope_for_nokogiri_matcher.should_not have_tag(tag, options, &block)
     end